# web/docu/swagger/swagger.yaml
basePath: /api/v1
definitions:
  config.AIConfig:
    properties:
      endpoint:
        type: string
      features:
        $ref: '#/definitions/config.AIFeatures'
      model:
        type: string
      provider:
        type: string
      selfHealing:
        $ref: '#/definitions/config.SelfHealing'
      token:
        type: string
    type: object
  config.AIFeatures:
    properties:
      anomalyDetection:
        type: boolean
      incidentAnalysis:
        type: boolean
      performanceOptimization:
        type: boolean
    type: object
  config.ClusterConfig:
    properties:
      brokers:
        type: string
      clusterID:
        type: string
      provider:
        type: string
      security:
        $ref: '#/definitions/config.SecurityConfig'
    type: object
  config.Config:
    properties:
      ai:
        $ref: '#/definitions/config.AIConfig'
      clusters:
        additionalProperties:
          $ref: '#/definitions/config.ClusterConfig'
        type: object
      database:
        $ref: '#/definitions/config.DatabaseConfig'
      logging:
        $ref: '#/definitions/config.LoggingConfig'
      monitoring:
        $ref: '#/definitions/config.MonitoringConfig'
      replication:
        $ref: '#/definitions/config.ReplicationConfig'
      server:
        $ref: '#/definitions/config.ServerConfig'
      topics:
        items:
          $ref: '#/definitions/config.TopicMapping'
        type: array
    type: object
  config.DatabaseConfig:
    properties:
      path:
        type: string
    type: object
  config.LoggingConfig:
    properties:
      console:
        type: boolean
      file:
        type: string
      level:
        type: string
      maxAge:
        type: integer
      maxBackups:
        type: integer
      maxSize:
        type: integer
    type: object
  config.LokiConfig:
    properties:
      endpoint:
        type: string
    type: object
  config.MonitoringConfig:
    properties:
      enabled:
        type: boolean
      loki:
        $ref: '#/definitions/config.LokiConfig'
      platform:
        description: '"splunk", "loki", "prometheus"'
        type: string
      prometheus:
        $ref: '#/definitions/config.PrometheusConfig'
      splunk:
        $ref: '#/definitions/config.SplunkConfig'
    type: object
  config.PrometheusConfig:
    properties:
      pushGateway:
        type: string
    type: object
  config.ReplicationConfig:
    properties:
      batchSize:
        type: integer
      compression:
        type: string
      jobID:
        type: string
      parallelism:
        type: integer
    type: object
  config.SecurityConfig:
    properties:
      apikey:
        type: string
      apisecret:
        type: string
      connectionString:
        type: string
      enabled:
        type: boolean
      kerberos:
        properties:
          serviceName:
            type: string
        type: object
      password:
        type: string
      protocol:
        type: string
      saslmechanism:
        type: string
      username:
        type: string
    type: object
  config.SelfHealing:
    properties:
      automatedRemediation:
        type: boolean
      dynamicThrottling:
        type: boolean
    type: object
  config.ServerConfig:
    properties:
      adminEmail:
        type: string
      cors:
        properties:
          allowedOrigins:
            items:
              type: string
            type: array
        type: object
      host:
        type: string
      mode:
        type: string
      port:
        type: integer
      tls:
        properties:
          certFile:
            type: string
          enabled:
            type: boolean
          keyFile:
            type: string
        type: object
      webPath:
        type: string
    type: object
  config.SplunkConfig:
    properties:
      hecendpoint:
        type: string
      hectoken:
        type: string
      index:
        type: string
    type: object
  config.TopicMapping:
    properties:
      enabled:
        type: boolean
      source:
        type: string
      target:
        type: string
    type: object
...
  database.KafkaCluster:
    properties:
      api_key:
        type: string
      api_secret:
        type: string
      brokers:
        type: string
      cluster_id:
        type: string
      connection_string:
        type: string
      name:
        type: string
      provider:
        type: string
      security_config:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  database.ReplicationJob:
    properties:
      batch_size:
        type: integer
      compression:
        type: string
      created_at:
        type: string
      failed_reason:
        type: string
      id:
        type: string
      name:
        type: string
      parallelism:
        type: integer
      preserve_partitions:
        type: boolean
      source_cluster_name:
        type: string
      status:
        type: string
      target_cluster_name:
        type: string
      updated_at:
        type: string
    type: object
...
  database.TopicMapping:
    properties:
      enabled:
        type: boolean
      id:
        type: integer
      job_id:
        type: string
      source_topic_pattern:
        type: string
      target_topic_pattern:
        type: string
    type: object
  database.User:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_initial:
        type: boolean
      username:
        type: string
    type: object
...
  server.CreateJobRequest:
    properties:
      batch_size:
        type: integer
      compression:
        type: string
      name:
        type: string
      parallelism:
        type: integer
      preserve_partitions:
        type: boolean
      source_cluster_name:
        type: string
      target_cluster_name:
        type: string
      topic_mappings:
        items:
          $ref: '#/definitions/database.TopicMapping'
        type: array
    type: object
...
  server.createUserRequest:
    properties:
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API for kaf-mirror, a high-performance Kafka replication
    tool.
  title: kaf-mirror API
  version: "1.0"
paths:
...
  /clusters:
    get:
      description: Get a list of all configured Kafka clusters.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.KafkaCluster'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List all Kafka clusters
      tags:
      - clusters
    post:
      consumes:
      - application/json
      description: Create a new Kafka cluster configuration.
      parameters:
      - description: Kafka Cluster
        in: body
        name: cluster
        required: true
        schema:
          $ref: '#/definitions/database.KafkaCluster'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/database.KafkaCluster'
      security:
      - ApiKeyAuth: []
      summary: Create a new Kafka cluster
      tags:
      - clusters
...
  /config:
    get:
      description: Get the current configuration of the kaf-mirror.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/config.Config'
      security:
      - ApiKeyAuth: []
      summary: Get the current configuration
      tags:
      - config
    put:
      consumes:
      - application/json
      description: Update the configuration of the kaf-mirror.
      parameters:
      - description: New configuration
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/config.Config'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update the configuration
      tags:
      - config
...
  /jobs:
    get:
      description: Get a list of all replication jobs.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.ReplicationJob'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List all replication jobs
      tags:
      - jobs
    post:
      consumes:
      - application/json
      description: Create a new replication job.
      parameters:
      - description: Replication Job
        in: body
        name: job
        required: true
        schema:
          $ref: '#/definitions/server.CreateJobRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/database.ReplicationJob'
      security:
      - ApiKeyAuth: []
      summary: Create a new replication job
      tags:
      - jobs
...
  /users:
    get:
      description: Get a list of all users.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - ApiKeyAuth: []
      summary: List all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user.
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/server.createUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/database.User'
      security:
      - ApiKeyAuth: []
      summary: Create a new user
      tags:
      - users
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"